requirements.txt
yt-dlp>=2024.1.0
requests>=2.31.0
browser-cookie3>=0.19.1
python-dotenv>=1.0.0
rich>=13.7.0
